Revert "remove hiveberg dependency in iceberg-spark2 module"
This reverts commit 2e8b7430ffd605b6b5e02f64304039def46f29c1.
