+++ build.gradle
@@ -498,15 +498,11 @@ project(':iceberg-hive-metastore') {
 project(':iceberg-hiveberg') {
   dependencies {
     compile project(':iceberg-hive-metastore')
-    compile project(':iceberg-spark2')
 
     compileOnly "org.apache.avro:avro"
     compileOnly("org.apache.hadoop:hadoop-client") {
       exclude group: 'org.apache.avro', module: 'avro'
       exclude group: 'org.slf4j', module: 'slf4j-log4j12'
-    }
-    compileOnly("org.apache.spark:spark-hive_2.11") {
-      exclude group: 'org.apache.avro', module: 'avro'
     }
 
     compileOnly("org.apache.hive:hive-metastore") {
@@ -876,6 +872,7 @@ if (jdkVersion == '8') {
       compile project(':iceberg-arrow')
       compile project(':iceberg-hive-metastore')
       compile project(':iceberg-spark')
+      compile project(':iceberg-hiveberg')
 
       compileOnly "org.apache.avro:avro"
       compileOnly("org.apache.spark:spark-hive_2.11") {
